refactor(go-template): add GoEmitContext seam; extract helper-inline + url-builder
Introduce `GoEmitContext` — the narrow interface (per-compile `state` plus the
recursive entry points `parseLiteralExpression` / `convertExpressionToGo` /
`convertConditionToGo`) that extracted emit modules depend on instead of the
concrete 8k-line adapter. The adapter implements it and passes `this`; modules
are free functions taking the context, so cross-module coupling is explicit and
the modules are testable against a stub.

First two clusters moved through the seam:
- expr/helper-inline.ts — local arrow-const helper inlining at a call site
  (inlineLocalHelperCall + the splice-safety guards + substituteHelperParams).
- expr/url-builder.ts — URLSearchParams builder helpers → bf_query (#1897).

Output byte-identical (unit 551, conformance 786).
